refactor: move decode_holocene_base_fee fn into separate basefee.rs
